+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Support for regular expressions to replace normalizer. See
+  `Tokenizers.Normalizer.replace_regex/2`.
+- Support for regular expressions to split pre-tokenizer. See
+  `Tokenizers.PreTokenizer.split_regex/3`.
+
 ## [v0.4.0] - 2023-08-09
 
 ### Added